requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.4            # via black, virtualenv
attrs==20.3.0             # via pytest
black==20.8b1             # via -r requirements.in
cfgv==3.2.0               # via pre-commit
click==7.1.2              # via black
distlib==0.3.1            # via virtualenv
filelock==3.0.12          # via virtualenv
identify==1.5.10          # via pre-commit
iniconfig==1.1.1          # via pytest
lark-parser==0.11.1       # via -r requirements.in
mypy-extensions==0.4.3    # via black
nodeenv==1.5.0            # via pre-commit
numpy==1.19.4             # via -r requirements.in
packaging==20.8           # via pytest
pathspec==0.8.1           # via black
pluggy==0.13.1            # via pytest
pre-commit==2.9.3         # via -r requirements.in
py==1.10.0                # via pytest
pyparsing==2.4.7          # via packaging
pytest==6.2.1             # via -r requirements.in
pyyaml==5.3.1             # via pre-commit
regex==2020.11.13         # via black
six==1.15.0               # via virtualenv
tabulate==0.8.7           # via -r requirements.in
toml==0.10.2              # via black, pre-commit, pytest
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
virtualenv==20.2.2        # via pre-commit
